Added better error handling and logging to dashboard.

+++ app/controllers/dashboard_controller.rb
@@ -8,14 +8,15 @@ class DashboardController < ApplicationController
     edit_user = edit_user?
     @page_title = "Dashboard"
     @summaries = []
+    @editors = editors()
 
-    EcoSummary.all.each do |s|
-      if s.public
-        @summaries << s
+    EcoSummary.all.each do |summary|
+      if summary.public == 1
+        @summaries << summary
         next
       end
-      if edit_user && s.created_by == safe_current_user
-        @summaries << s
+      if edit_user && summary.created_by == safe_current_user
+        @summaries << summary
         next
       end
     end
@@ -26,13 +27,20 @@ class DashboardController < ApplicationController
     else
       "https://search.library.brown.edu/users/auth/shibboleth?target=" + dashboard_new_url()
     end
+
+    Rails.logger.info("Show dashboard list for user #{safe_current_user}.")
     render
   end
 
   def show
-    @summaries = EcoSummary.all
     id = (params["id"] || 0).to_i
     summary = EcoSummary.find(id)
+    if summary.public != 1 && summary.created_by != safe_current_user
+      Rails.logger.error("User #{safe_current_user} does not have access to summary #{id}")
+      render "error", status: 500
+      return
+    end
+    Rails.logger.info("Show dashboard #{id} for user #{safe_current_user}.")
     @page_title = summary.list_name
     @presenter = DashboardDetailsPresenter.new(summary)
     @presenter.download_url = dashboard_details_url(id: id, format: 'tsv')
@@ -45,10 +53,20 @@ class DashboardController < ApplicationController
 
   def edit
     if !edit_user?
-        raise "Invalid user: #{safe_current_user}."
+        Rails.logger.error("User #{safe_current_user} does not have access to edit")
+        render "error", status: 500
+        return
     end
     id = (params["id"] || 0).to_i
     summary = EcoSummary.find(id)
+
+    if summary.public != 1 && summary.created_by != safe_current_user
+      Rails.logger.error("User #{safe_current_user} does not have access to edit summary #{id}")
+      render "error", status: 500
+      return
+    end
+
+    Rails.logger.info("Edit dashboard #{id} for user #{safe_current_user}.")
     @presenter = DashboardDetailsPresenter.new(summary)
     @presenter.edit_user = true
     @page_title = summary.list_name
@@ -60,9 +78,20 @@ class DashboardController < ApplicationController
 
   def delete
     if !edit_user?
-      raise "Invalid user: #{safe_current_user}."
+      Rails.logger.error("User #{safe_current_user} does not have access to delete")
+      render status: 400, :json => "{\"message\": \"User does not have access to delete\"}"
+      return
     end
+
     id = (params["id"] || 0).to_i
+    summary = EcoSummary.find(id)
+    if summary.public != 1 && summary.created_by != safe_current_user
+      Rails.logger.error("User #{safe_current_user} does not have access to delete summary #{id}")
+      render status: 400, :json => "{\"message\": \"User does not have access to delete this summary\"}"
+      return
+    end
+
+    Rails.logger.info("Delete dashboard #{id} by user #{safe_current_user}.")
     EcoSummary.delete(id)
     EcoRange.delete_all(eco_summary_id: id)
     # TODO: delete details
@@ -70,14 +99,18 @@ class DashboardController < ApplicationController
   end
 
   def details
-    @summaries = EcoSummary.all
-
     id = (params["id"] || 0).to_i
     all = (params["all"] == "yes")
     max = all ? -1 : 1000
     rows = []
     count = 0
     summary = EcoSummary.find(id)
+    if summary.public != 1 && summary.created_by != safe_current_user
+      Rails.logger.error("User #{safe_current_user} does not have access to details for summary #{id}")
+      render "error", status: 500
+      return
+    end
+
     @page_title = summary.list_name
     @presenter = nil
 
@@ -93,28 +126,32 @@ class DashboardController < ApplicationController
       @presenter.to = range.to
       @presenter.download_url = dashboard_details_url(id: id, range_id: range_id, all: 'yes', format: 'tsv')
       @presenter.show_all_url = dashboard_details_url(id: id, range_id: range_id, all: 'yes')
+      Rails.logger.info("Show details (range) for dashboard #{id} for user #{safe_current_user}.")
     elsif loc_code != ""
       count, rows = EcoDetails.by_location(loc_code, summary.id, max)
       name = "Location #{Location.get_name(loc_code)} (#{loc_code})"
       @presenter = DashboardDetailsPresenter.new(summary, name, count, rows)
       @presenter.download_url = dashboard_details_url(id: id, loc_code: loc_code, all: 'yes', format: 'tsv')
       @presenter.show_all_url = dashboard_details_url(id: id, loc_code: loc_code, all: 'yes')
+      Rails.logger.info("Show details (location) for dashboard #{id} for user #{safe_current_user}.")
     elsif ck_count > 0
       count, rows = EcoDetails.by_usage(ck_count, summary.id, max)
       name = "Checked out #{ck_count} times"
       @presenter = DashboardDetailsPresenter.new(summary, name, count, rows)
       @presenter.download_url = dashboard_details_url(id: id, ck: ck_count, all: 'yes', format: 'tsv')
       @presenter.show_all_url = dashboard_details_url(id: id, ck: ck_count, all: 'yes')
+      Rails.logger.info("Show details (usage) for dashboard #{id} for user #{safe_current_user}.")
     else
       count, rows = EcoDetails.by_summary(summary.id, max)
       name = max == -1 ? "All items" : "First #{max} items"
       @presenter = DashboardDetailsPresenter.new(summary, name, count, rows)
       @presenter.download_url = dashboard_details_url(id: id, all: 'yes', format: 'tsv')
       @presenter.show_all_url = dashboard_details_url(id: id, all: 'yes')
+      Rails.logger.info("Show details (all) for dashboard #{id} for user #{safe_current_user}.")
     end
 
     if params["format"] == "tsv"
-      Rails.logger.info("Exporting TSV: #{@presenter.summary.list_name} - #{@presenter.name}, #{@presenter.count} records")
+      Rails.logger.info("Exporting TSV: #{@presenter.summary.list_name} - #{@presenter.name}, #{@presenter.count} records for user #{safe_current_user}")
       send_data(EcoDetails.to_tsv(@presenter.rows), :filename => "dashboard_#{summary.id}.tsv", :type => "text/tsv")
       return
     end
@@ -128,7 +165,9 @@ class DashboardController < ApplicationController
 
   def new
     if !edit_user?
-      raise "Invalid user: #{safe_current_user}."
+      Rails.logger.error("User #{safe_current_user} does not have access to create new")
+      render "error", status: 500
+      return
     end
     summary = EcoSummary.new()
     summary.list_name = "#{safe_current_user}'s new list"
@@ -138,25 +177,39 @@ class DashboardController < ApplicationController
     summary.created_by = safe_current_user
     summary.public = 0
     summary.save
+    Rails.logger.info("Created dashboard #{summary.id} for user #{safe_current_user}.")
     redirect_to dashboard_edit_url(id: summary.id)
   end
 
   def copy
     if !edit_user?
-      raise "Invalid user: #{safe_current_user}."
+      Rails.logger.error("User #{safe_current_user} does not have access to create copy")
+      render "error", status: 500
+      return
     end
 
     id = (params["id"] || 0).to_i
-    summary = EcoSummary.copy(id, current_user)
+    summary = EcoSummary.copy(id, safe_current_user)
+    Rails.logger.info("Created dashboard #{summary.id} from #{id} for user #{safe_current_user}.")
     redirect_to dashboard_edit_url(id: summary.id)
   end
 
   def save
     if !edit_user?
-      raise "Invalid user: #{safe_current_user}."
+      Rails.logger.error("User #{safe_current_user} does not have access to save")
+      render "error", status: 500
+      return
     end
     id = (params["id"] || 0).to_i
     summary = EcoSummary.find(id)
+
+    if summary.public != 1 && summary.created_by != safe_current_user
+      Rails.logger.error("User #{safe_current_user} does not have access to save summary #{id}")
+      render "error", status: 500
+      return
+    end
+
+    Rails.logger.info("Saved dashboard #{summary.id} for user #{safe_current_user}.")
     summary.save_from_request(params, safe_current_user)
     redirect_to dashboard_show_url(id: id)
   rescue ActiveRecord::RecordNotFound
@@ -165,12 +218,23 @@ class DashboardController < ApplicationController
   end
 
   private
+    def editors()
+      return [] if !edit_user?
+      data = []
+      (ENV["DASHBOARD_USERS"] || "").split("/").each do |user|
+        if (user != "")
+          data << user
+        end
+      end
+      data
+    end
+
     def safe_current_user
       if ENV["LOCALHOST"] == "true"
         return "josiah@localhost"
+      end
+      current_user
     end
-    current_user
-  end
 
     def edit_user?
       if ENV["LOCALHOST"] == "true"
